models: Change the Release table name
Turns out that 'user' is a reserved keyword on many SQL platforms,
so we can't use that for our User table.

I used 'users' to work around that, but then it really is an eye sore
(at least to me) to have one table name singular and one plural.

So let's make this one plural too, and use plural table names for the
rest in the future.