js: Rename 'release' to 'item'
This commit does not make the code completely generic yet, it
still assumes in many places that we are dealing with releases
only.

However, it is a trivial rename which will help on making the
next few commits easier to understand, as there won't be so
many changes in them.