config: Declare the views generically
The views were generic, but we still declared them as relative to the
release management.

With this commit, all we will have to do is add a new resource type for
it to have a management interface.

This was the last step in making the management code generic and
reusable!