admin: Use deform for the admin form
So, it turns out that I missed something when I renamed the release
template to admin.pt. It wasn't completely generic yet: the form was
still very much tied to the release and its attributes. :)

This commit brings in the excellent deform_bootstrap, which allows us to
use deform to generate our forms while integrating in our Bootstrap
theme with basically no change.