overview: Finally get rid of that ugly grouping
The template code is a bit more complex, but it just feels much better
that the view doesn't "know" how many release per line the template
wants to display.

I'm leaving the view code with the loop, which is more complex than
needed, but will minimize differences with the future code.