distros: Add an optional relation to upstreams
In some cases, all the packages in a distro will have the same upstream.
For these, it makes sense to just have the upstream be an attribute of
the distro.

However let's keep it optional, for these other cases where each package
of a given distro can come from a different upstream, as is the case for
us at Network Box.