utils: Drop the (now useless) version comparisons
The sync code is now the only user of these, and all it needs to check
is whether EVRS are strictly identical or not.

The naive string comparison works well for this, which is why we keep
__eq__ and __ne__.

However, checking inferiority/superiority really needs the RPM API, so
the __lt__, __le__, __gt__ and __ge__ can not be correct in any case,
which warrants their removal to avoid future mistakes.