distros: Add a downstream_prefix attribute
When changing the packages downstream, it is a good idea to mark such
changes by modifying the Release: tag in the spec file.

But just bumping the release doesn't work, as there is no way to know in
advance what will be the next release of the package from upstream, and
as such we risk breaking the upgrade path.

The best way, thus, is to append something to the Release tag from
upstream.

The people at GoOSe Linux append '.goose1' on their first change and
then increase the number at every subsequent change.

At Network Box, we just use '.1', '.2', etc.

In both case, the easiest way to handle this when comparing versions
will just be to search for the "downstream prefix" (in the examples
above, either '.goose' or '.') and then ignore what comes after it.

This commit makes it possible to specify the convention used for each
distro.