dedist: Sort dist tags before removing them
Here at Network Box, we have an "EL6" upstream, which can have the
following dist tags:
    ".el6", ".el6_0", ".el6_1", ".el6_2"

The previous dedist code would take them in whatever order they were
defined, which happened to be the order shown above.

As such, it would take an EVR, say "1.0-1.el6_2" and try removing the
dists one after the other.

So it would try removing the first one (".el6"), which gave as a result:
    "1.0-1_2"

And then it would try removing the other ones, but none would match so
the above line would be the final EVR.

But that was wrong, the final EVR should have been:
    "1.0-1"

I.e the dist tag which should have been removed was ".el6_2", not
".el6".

Sorting them (most recent first) fixes this issue.