sync: Get the upstream for a package from Git
In some cases (as is the case for us at Network Box), all the packages
in a given Distro do not come from the same Upstream.

We then need a way to find the Upstream for each Package.

The way we chose to follow here at Network Box is to respect a certain
naming convention in our Git branches.

As a result, we can identify which was the latest upstream branch merged
in the branch for the current Distro.