sync: Cache the latest Package for each Upstream
Very often, more than one Distro will share the same Upstream for a
given Package.

In these cases, it is ok to consider that Upstream hasn't published a
new version of the Package while we were synchronizing.

This assumption allows speeding up the synchronization script quite a
lot (from 55 to 35 minutes here!).