overview: Don't display upstreams with both counts at 0
If all the packages coming from a certain Upstream are marked as
problematic, then the counts for that upstream will be at 0, both for up
to date and out of date packages.

As a result, it shouldn't appear in the overview, to avoid cluttering
the space and confusing the viewers.