status: Don't split table labels on two lines
When the EVR is very long, the table try to keep its width and as a
result we'd get things like:

    Upstream
    EVR

Which just looks plain ugly.

And if the EVR was **really** very long, then the table width would grow
anyway.

Using unbreakable spaces means the table will grow earlier, but we won't
get the ugly line-breaking.