gitbase: Keep the cloned dirs around
This should speed up the upstream branch retrieval quite a lot, as it
means we won't have to clone every single time.

It is a bit dangerous though, as anything else than uptrack could modify
the cloned trees. :-/

If that turns out to be the wrong tradeoff between speed and security,
we'll revert this change.