sync: Don't catch all errors
There is one type of error we know about and want to deal specifically
(by logging them and processing the package accordingly) : when we can't
find the upstream of a package.

But there might be other errors (and in fact, there is still a NameError
in the get_upstream function), and these should not be merely logged and
glossed over.